Fix compiler warning
git-svn-id: https://svn.osgeo.org/gdal/trunk/gdal@22873 f0d54148-0727-0410-94bb-9a71ac55c965

+++ ogr/ogrsf_frmts/vfk/vfkdatablock.cpp
@@ -890,7 +890,7 @@ long VFKDataBlock::LoadGeometry()
     end = clock();
     
     CPLDebug("OGR_VFK", "VFKDataBlock::LoadGeometry(): name=%s nfeatures=%ld sec=%ld",
-	     m_pszName, nfeatures, (end - start) / CLOCKS_PER_SEC);
+	     m_pszName, nfeatures, (long)((end - start) / CLOCKS_PER_SEC));
     
     return nfeatures;
 }
